requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
pydantic==1.10.2
python-decouple==3.6
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.20.0
